refactor: split ngx_keyval.c into separate files by responsibility
Split the monolithic ngx_keyval.c (990 lines) into focused files:
- ngx_keyval_zone.c: zone management (add, get, set_zone, set_zone_redis)
- ngx_keyval_conf.c: config parsing (set_variable, create_main_conf)
- ngx_keyval_variable.c: runtime key construction (variable_get_key)
- ngx_keyval_store_shm.c: shared memory storage (rbtree, init, get/set)
- ngx_keyval_store_redis.c: Redis storage operations

No logic changes - pure code movement with visibility adjustments
for functions that need cross-file access.
